I realized that I had to get rid of the functions and code that was making the ship face up instead of how I wanted in the center.
To avoid chaos. I cloned my original repo, made this new one, copied my files into this new repo and reworked my code from the previous repo, got rid of the up and down functions and made the rocket ship face the correct way as it would spawn on the center of the right side. I also corrected the bullets and this game is ready for future development.